% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.r
\name{auc}
\alias{auc}
\title{Estimates the area under the curve (auc)
for a smoothed PBI time series. Data are
normalized to allow for comparisons of rate
of change.}
\usage{
auc(df, span = 0.3)
}
\arguments{
\item{df:}{cropmonitor data frame / database file}

\item{span:}{smoothing factor for loess fitting}
}
\description{
Estimates the area under the curve (auc)
for a smoothed PBI time series. Data are
normalized to allow for comparisons of rate
of change.
}
\examples{
# no examples yet
}
\keyword{index}
\keyword{series,}
\keyword{time}
