% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_glcm.r
\name{calculate_glcm}
\alias{calculate_glcm}
\title{Estimates glcm values for a region of interest}
\usage{
calculate_glcm(img, roi = NULL)
}
\arguments{
\item{img:}{RGB image to process (filename or 3-layer RGB stack or brick)}

\item{plot:}{plot resulting image with all available information}
}
\description{
Estimates glcm values for a region of interest
}
\examples{
# no examples yet
}
\keyword{entropy}
\keyword{feature,}
\keyword{glcm,}
