% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sobel.r
\name{calculate_sobel}
\alias{calculate_sobel}
\title{Estimates the gcc value for a region of interest
automatically generated by any given algorithm}
\usage{
calculate_sobel(img, roi = NULL)
}
\arguments{
\item{img:}{RGB image to process (filename or 3-layer RGB stack or brick)}

\item{plot:}{plot resulting image with all available information}
}
\description{
Estimates the gcc value for a region of interest
automatically generated by any given algorithm
}
\examples{
# no examples yet
}
\keyword{detection,}
\keyword{edge}
\keyword{feature}
\keyword{index,}
\keyword{sobel}
