% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropmonitor.r
\name{cropmonitor}
\alias{cropmonitor}
\title{Start the cropmonitor shiny interface}
\usage{
cropmonitor(path = "~/cropmonitor", rdsfile = "cropmonitor.rds")
}
\arguments{
\item{path}{: a path were the data is stored (default = ~/cropmonitor/)}
}
\description{
Start the cropmonitor shiny interface
}
\examples{

## NOT RUN

cropmonitor()

## END NOT RUN
}
\keyword{GUI}
