% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_horizon.r
\name{estimate_horizon}
\alias{estimate_horizon}
\title{Start the cropmonitor shiny interface}
\usage{
estimate_horizon(img, penalty = 0.2, plot = FALSE)
}
\arguments{
\item{img:}{single raster layer to process}

\item{penalty:}{parameter to set the sensitivity of the changepoint
detection used in finding the horizon.}

\item{plot:}{visualize the horizon on the image}
}
\description{
Start the cropmonitor shiny interface
}
\examples{
# no examples yet
}
\keyword{interest}
\keyword{of}
\keyword{region}
\keyword{selection}
