% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_roi.r
\name{estimate_roi}
\alias{estimate_roi}
\title{Estimates a region of interest (ROI) using the
approximate horizon location in an image.}
\usage{
estimate_roi(img, padding = 0.1, plot = FALSE)
}
\arguments{
\item{img:}{RGB image to process (filename or 3-layer RGB stack or brick)}

\item{padding:}{% of the image width / height to pad}

\item{plot:}{plot resulting image with all available information}
}
\description{
This method works auto-magically but might
fail in some instances. As such there is a
plotting and output function to visualize the
actual ROI for quality control
}
\examples{
# no examples yet
}
\keyword{interest}
\keyword{of}
\keyword{region}
\keyword{selection}
