% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Folder2Label.r
\name{Folder2Label}
\alias{Folder2Label}
\title{Takes label folders for tensor training and stores labels  in the image database}
\usage{
Folder2Label(database, image_path, image_ext = ".jpg", out_image_path)
}
\arguments{
\item{database:}{path and name of RDS or dta file as provided by IFPRI with images where labels will be stored}

\item{image_ext:}{extension type of image for example .jpg}

\item{out_image_path:}{path of folder where images were sorted for training}
}
\description{
Files should be sorted in the following order:
out_image_path:
      LabelFolder1:
           TrueFolder:
               Image_i.jpg
           FalseFolder:
               Image_i.jpg
           OtherFolder:
     LabelFolder2:
           TrueFolder:
               Image_i.jpg
           FalseFolder:
           OtherFolder:
}
\examples{

\dontrun{
image_ext = '.jpg'
database = '/media/ssd/Lodging_Classifier/cropmonitor_merged.rds'
out_image_path = '/media/ssd/Lodging_Classifier/sorted'
in_data = Folder2Label(database,image_path,image_ext='.jpg',out_image_path)
saveRDS(in_data,'/media/ssd/Lodging_Classifier/cropmonitor_merged_updated_images.rds')

table(in_data$Lodging)
}
}
\keyword{label}
\keyword{sort}
