% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label2Folder.r
\name{Label2Folder}
\alias{Label2Folder}
\title{Takes labels and moves associated files to folders for tensor training}
\usage{
Label2Folder(database, image_path, image_ext = ".jpg", out_image_path)
}
\arguments{
\item{database:}{path and name of RDS or dta file as provided by IFPRI with images where labels will be stored}

\item{database_label:}{column of database where labels are stored}

\item{database_image_col_name:}{column of database where names, urls or paths to images are stored}

\item{image_path:}{path of folder holding aoi images}

\item{image_ext:}{extension type of image for example .jpg}

\item{out_image_path:}{path of folder where images should be sorted for training}
}
\description{
Takes labels and moves associated files to folders for tensor training
}
\examples{

\dontrun{
image_path = '/media/ssd/Lodging_Classifier/aoi/'
image_ext = '.jpg'
database = '/media/ssd/Lodging_Classifier/cropmonitor_merged.rds'
database_label = 'Lodging'
database_image_col_name = 'image'
out_image_path = '/media/ssd/Lodging_Classifier/sorted'
Label2Folder(database,image_path,image_ext='.jpg',out_image_path)
}
}
