% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cropmonitor_sites.r
\name{plot_cropmonitor_sites}
\alias{plot_cropmonitor_sites}
\title{Plot Gcc and GRVI values for all fields (sites), for quick
non interactive review or reporting}
\usage{
plot_cropmonitor_sites(database = "~/cropmonitor/cropmonitor.json",
  questionaire = "~/cropmonitor/questionaire.xlsx", span = 0.3,
  out_dir = "~/cropmonitor")
}
\arguments{
\item{database:}{json clean database file}

\item{span:}{smoothing parameter (% of data points)}

\item{out_dir:}{path where to store the output pdfs}
}
\description{
Plot Gcc and GRVI values for all fields (sites), for quick
non interactive review or reporting
}
\examples{


}
\keyword{/}
\keyword{QA/GC}
\keyword{calculation,}
\keyword{gcc}
\keyword{grvi}
