% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_location.r
\name{plot_spatial_location}
\alias{plot_spatial_location}
\title{Plot the spatial location of a PBI farmer
to determine if the geographic location
of the time series corresponds with a field
on the ground. Used in QA/QC for satellite
data comparisons.}
\usage{
plot_spatial_location(df = df, userfield = NULL, path = "~",
  stack_path = "~")
}
\arguments{
\item{df}{cropmonitor database dataframe}

\item{userfield}{userfield, concated userid and field # as in user-field}

\item{path}{path location where to output summary graphics}

\item{stack_path}{path to subsets of the HLS images, to map the grid of the
harmonized landsat sentinel data}
}
\description{
Plot the spatial location of a PBI farmer
to determine if the geographic location
of the time series corresponds with a field
on the ground. Used in QA/QC for satellite
data comparisons.
}
\examples{
# no examples yet
\dontrun{
# Load main data file, including all indices
df = readRDS("~/cropmonitor/cropmonitor.rds")

# screen for test sites in the US / EU
df = df[which(df$longitude > 70),]
df = df[which(as.Date(df$date) > as.Date("2016-10-01")),]

# create unique field vector
df$userfield = paste(df$uniqueuserid,df$uniquecropsiteid,sep = "-")

# plot spatial location graph in home directory
plot_spatial_location(df, userfield = 430720-1, path = '~')

}
}
\keyword{QA/QC}
