% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_size.r
\name{read_size}
\alias{read_size}
\title{Reads and resizes images or raster layers
rotates it in the proper direction when necessary}
\usage{
read_size(img, ncols = 640, nrows = NULL)
}
\arguments{
\item{img:}{RGB image, or raster file to process}

\item{ncols:}{resize to a particular number of columns}

\item{nrows:}{resize to a particular number of rows,
if NULL the same ratio is used as set by the original ncol
vs. the specified ncols.}
}
\description{
Reads and resizes images or raster layers
rotates it in the proper direction when necessary
}
\keyword{image}
\keyword{preparation,}
\keyword{reading}
