% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_farmer_reports.r
\name{render_farmer_reports}
\alias{render_farmer_reports}
\title{Plot farmer reports with damages (non GUI version)}
\usage{
render_farmer_reports(database = "~/cropmonitor/cropmonitor.rds",
  payout = "~/cropmonitor/payouts_rabi1617.dta", out_dir = "~")
}
\arguments{
\item{database:}{json clean database file (cropmonitor.rds)}

\item{payout:}{path to payout dta file}

\item{out_dir:}{path where to store the output pdfs}
}
\description{
Plot farmer reports with damages (non GUI version)
}
\keyword{reporting}
