% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_glcm.r
\name{calculate_glcm}
\alias{calculate_glcm}
\title{Estimates the gcc value for a region of interest
automatically generated by any given algorithm}
\usage{
calculate_glcm(img, roi = NULL, plot = FALSE)
}
\arguments{
\item{img:}{RGB image to process (filename or 3-layer RGB stack or brick)}

\item{plot:}{plot resulting image with all available information}
}
\description{
Estimates the gcc value for a region of interest
automatically generated by any given algorithm
}
\examples{
# no examples yet
}
\keyword{calculation}
\keyword{gcc}
