% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_horizon.r
\name{estimate_horizon}
\alias{estimate_horizon}
\title{Start the cropmonitor shiny interface}
\usage{
estimate_horizon(img, plot = FALSE)
}
\arguments{
\item{img:}{image to process}
}
\description{
Start the cropmonitor shiny interface
}
\examples{
# no examples yet
}
\keyword{interest}
\keyword{of}
\keyword{region}
\keyword{selection}
