% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_image_db.r
\name{process_image_db}
\alias{process_image_db}
\title{Batch process all data in the IFPRI database using the
automatic ROI detection methodology}
\usage{
process_image_db(database = NULL, path = "~/cropmonitor/",
  server = "http://cdn.wheatcam.ifpri.org/ReportImages", thumbnails = FALSE,
  force = TRUE, ncores = 6)
}
\arguments{
\item{database:}{STATA file as provided by IFPRI}

\item{path:}{path of the IFPRI database images to process}

\item{plot:}{TRUE / FALSE (output summary plots?)}

\item{force:}{force regeneration of all indices}

\item{force_all:}{force regeneration of all data in database
including estimating the horizon and ROI, as well as indices / features}
}
\description{
Batch process all data in the IFPRI database using the
automatic ROI detection methodology
}
\examples{
# no examples yet
}
\keyword{QA/GC}
\keyword{calculation,}
\keyword{gcc}
